refactor(timer_alert): split TimerAlertFrame into mixin package
Convert widgets/frames/timer_alert.py (640 LOC) into a package with
focused modules, matching the match_history layout established in the
prior commit:

- __init__.py: public API (TimerAlertFrame, main) plus re-exports of
  mtgo_bridge, wx, and threading so existing headless tests can still
  ``monkeypatch.setattr(timer_alert.<module>, ...)`` on the same module
  objects the package references internally
- frame.py: ThresholdPanel helper, TimerAlertFrame.__init__ / _build_ui,
  panel-management and styling helpers, main()
- handlers.py: TimerAlertHandlersMixin — monitor start/stop, timer-step
  driver, alert playback, bridge-watch worker chain
  (_start_watch_loop, _watch_start_worker, _complete_watch_start,
  _stop_watcher_async, _stop_watcher_worker), timer/resize events,
  on_close
- properties.py: TimerAlertPropertiesMixin — translation (_t),
  status updates, _parse_thresholds, _format_seconds,
  _update_challenge_display, _challenge_wrap_width

Composition becomes:
  class TimerAlertFrame(HandlersMixin, PropertiesMixin, wx.Frame): ...

The tests in tests/test_timer_alert_async.py continue to patch
``timer_alert.mtgo_bridge.start_watch``, ``timer_alert.wx.CallAfter``,
and ``timer_alert.threading.Thread``; because those attributes are now
exposed via ``__init__.py`` and point to the same shared module objects
imported inside ``handlers.py``, patches still reach the runtime
call sites.

pyproject.toml's E402 per-file-ignore now points at the new
widgets/frames/timer_alert/frame.py where the sys.path bootstrap lives.

Refs #427 (second pass, partial).
