+++ pyproject.toml
@@ -70,7 +70,7 @@ ignore = [
 # Widget entry points insert project root into sys.path before other imports
 "widgets/frames/identify_opponent.py" = ["E402"]
 "widgets/frames/match_history/frame.py" = ["E402"]
-"widgets/frames/timer_alert.py" = ["E402"]
+"widgets/frames/timer_alert/frame.py" = ["E402"]
 "widgets/frames/metagame_analysis.py" = ["E402"]
 
 [tool.black]
