+++ widgets/frames/timer_alert/frame.py
@@ -1,22 +1,20 @@
-"""wxPython variant of the MTGO challenge timer alert widget."""
+"""UI construction for the timer alert widget."""
 
 from __future__ import annotations
 
 import sys
 from pathlib import Path
 
-_project_root = Path(__file__).resolve().parent.parent.parent
+_project_root = Path(__file__).resolve().parent.parent.parent.parent
 if str(_project_root) not in sys.path:
     sys.path.insert(0, str(_project_root))
 
 import re
-import threading
 from typing import Any
 
 import wx
 from loguru import logger
 
-from utils import mtgo_bridge
 from utils.constants import (
     DARK_ACCENT,
     DARK_ALT,
@@ -46,14 +44,8 @@ from utils.constants import (
 )
 from utils.i18n import translate
 from utils.mtgo_bridge_client import BridgeWatcher
-
-try:
-    import winsound  # type: ignore[attr-defined]
-
-    SOUND_AVAILABLE = True
-except Exception:  # pragma: no cover - fallback for non-Windows environments
-    SOUND_AVAILABLE = False
-    logger.warning("winsound not available, alarm sounds will not play")
+from widgets.frames.timer_alert.handlers import TimerAlertHandlersMixin
+from widgets.frames.timer_alert.properties import TimerAlertPropertiesMixin
 
 # Built-in Windows sounds (always available)
 SOUND_OPTIONS = {
@@ -117,7 +109,7 @@ class ThresholdPanel(wx.Panel):
         self.remove_btn.Enable(enabled)
 
 
-class TimerAlertFrame(wx.Frame):
+class TimerAlertFrame(TimerAlertHandlersMixin, TimerAlertPropertiesMixin, wx.Frame):
     """Polls MTGO challenge timers via the bridge and plays audible alerts."""
 
     WATCH_INTERVAL_MS = TIMER_ALERT_WATCH_INTERVAL_MS
@@ -161,10 +153,7 @@ class TimerAlertFrame(wx.Frame):
 
         wx.CallAfter(self._start_watch_loop)
 
-    def _t(self, key: str, **kwargs: object) -> str:
-        return translate(self._locale, key, **kwargs)
-
-    # ------------------------------------------------------------------ UI ------------------------------------------------------------------
+    # ------------------------------------------------------------------ UI build
     def _build_ui(self) -> None:
         self.SetBackgroundColour(DARK_BG)
 
@@ -194,10 +183,8 @@ class TimerAlertFrame(wx.Frame):
         self.threshold_container.SetSizer(self.threshold_container_sizer)
         threshold_sizer.Add(self.threshold_container, 1, wx.EXPAND | wx.ALL, PADDING_SM)
 
-        # Add initial threshold
         self._add_threshold_panel()
 
-        # Add threshold button
         add_btn = wx.Button(box_parent, label="+ Add Another Threshold")
         self._stylize_secondary_button(add_btn)
         add_btn.Bind(wx.EVT_BUTTON, lambda _evt: self._add_threshold_panel())
@@ -209,7 +196,6 @@ class TimerAlertFrame(wx.Frame):
         options_grid = wx.FlexGridSizer(cols=2, hgap=PADDING_BASE, vgap=PADDING_BASE)
         options_grid.AddGrowableCol(1, 1)
 
-        # Sound selection
         options_grid.Add(
             self._static_text(panel, self._t("timer.label.sound")), 0, wx.ALIGN_CENTER_VERTICAL
         )
@@ -218,7 +204,6 @@ class TimerAlertFrame(wx.Frame):
         self.sound_choice.SetSelection(0)
         options_grid.Add(self.sound_choice, 0, wx.EXPAND)
 
-        # Poll interval
         options_grid.Add(
             self._static_text(panel, self._t("timer.label.check_interval")),
             0,
@@ -233,7 +218,6 @@ class TimerAlertFrame(wx.Frame):
         self._stylize_spin(self.poll_interval_ctrl)
         options_grid.Add(self.poll_interval_ctrl, 0, wx.EXPAND)
 
-        # Repeat interval
         options_grid.Add(
             self._static_text(panel, self._t("timer.label.repeat_interval")),
             0,
@@ -352,273 +336,6 @@ class TimerAlertFrame(wx.Frame):
         font.MakeBold()
         button.SetFont(font)
 
-    # ------------------------------------------------------------------ Monitoring ------------------------------------------------------------
-    def start_monitoring(self) -> None:
-        if self.monitor_job_active:
-            self._set_status("timer.status.already_monitoring")
-            return
-
-        thresholds = self._parse_thresholds()
-        if not thresholds:
-            self._set_status("timer.status.no_thresholds")
-            return
-
-        try:
-            poll_interval = max(250, int(self.poll_interval_ctrl.GetValue()))
-        except (TypeError, ValueError):
-            self._set_status("timer.status.invalid_poll_interval")
-            return
-
-        self._current_thresholds = thresholds
-        self._monitor_interval_ms = poll_interval
-        self._repeat_interval_ms = self.repeat_interval_ctrl.GetValue() * 1000
-        self.triggered_thresholds.clear()
-        self.start_alert_sent = False
-        self.monitor_job_active = True
-
-        # Disable threshold editing
-        for panel in self.threshold_panels:
-            panel.set_enabled(False)
-
-        self._monitor_timer.Start(self._monitor_interval_ms)
-        self._set_status("timer.status.monitoring")
-        self._monitor_timer_step()
-
-    def stop_monitoring(self) -> None:
-        if self._monitor_timer.IsRunning():
-            self._monitor_timer.Stop()
-        if self._repeat_timer.IsRunning():
-            self._repeat_timer.Stop()
-        self.monitor_job_active = False
-
-        # Re-enable threshold editing
-        for panel in self.threshold_panels:
-            panel.set_enabled(True)
-
-        self._set_status("timer.status.stopped")
-
-    def test_alert(self) -> None:
-        self._play_alert()
-
-    def _parse_thresholds(self) -> list[int]:
-        thresholds: list[int] = []
-        for panel in self.threshold_panels:
-            seconds = panel.get_seconds()
-            if seconds is not None and seconds > 0:
-                thresholds.append(seconds)
-            elif panel.time_input.GetValue().strip():
-                logger.warning(f"Invalid threshold format: {panel.time_input.GetValue()}")
-        thresholds.sort(reverse=True)
-        return thresholds
-
-    def _monitor_timer_step(self) -> None:
-        snapshot = self._last_snapshot
-        if snapshot is None:
-            self._set_status("timer.status.waiting_data")
-            return
-
-        if snapshot.get("error"):
-            self._set_status("timer.status.bridge_error", error=snapshot["error"])
-            return
-
-        self._update_challenge_display(snapshot)
-
-        # Get challenge timer info
-        timers = snapshot.get("challengeTimers") or []
-        if not timers:
-            self._set_status("timer.status.no_timer")
-            self.triggered_thresholds.clear()
-            self.start_alert_sent = False
-            if self._repeat_timer.IsRunning():
-                self._repeat_timer.Stop()
-            return
-
-        timer = timers[0]
-        remaining = timer.get("remainingSeconds")
-        if not isinstance(remaining, (int, float)):
-            self._set_status("timer.status.invalid_value")
-            return
-
-        current_seconds = max(0, int(remaining))
-        self._set_status(
-            "timer.status.challenge_timer", value=self._format_seconds(current_seconds)
-        )
-
-        # Start alert (countdown began)
-        if self.start_alert_checkbox.GetValue() and not self.start_alert_sent:
-            self._trigger_alert("Countdown started")
-            self.start_alert_sent = True
-            if self.repeat_alarm_checkbox.GetValue():
-                self._repeat_timer.Start(self._repeat_interval_ms)
-
-        # Threshold alerts
-        for threshold in self._current_thresholds:
-            if threshold < 0 or threshold in self.triggered_thresholds:
-                continue
-            if current_seconds <= threshold:
-                self._trigger_alert(f"Timer reached {self._format_seconds(threshold)}")
-                self.triggered_thresholds.add(threshold)
-
-    def _trigger_alert(self, message: str) -> None:
-        logger.debug(f"Timer alert: {message}")
-        self._play_alert()
-
-    def _play_alert(self) -> None:
-        if not SOUND_AVAILABLE:
-            logger.warning("Sound playback not available")
-            return
-
-        sound_name = self.sound_choice.GetStringSelection()
-        sound_key = SOUND_OPTIONS.get(sound_name, "SystemDefault")
-
-        try:
-            winsound.PlaySound(sound_key, winsound.SND_ALIAS | winsound.SND_ASYNC)
-        except Exception as exc:  # noqa: BLE001
-            logger.debug(f"Failed to play sound: {exc}")
-
-    def _on_repeat_timer(self, _event: wx.TimerEvent) -> None:
-        if self.monitor_job_active and self.repeat_alarm_checkbox.GetValue():
-            self._play_alert()
-
-    def _format_seconds(self, value: Any) -> str:
-        if not isinstance(value, (int, float)):
-            return "—"
-        total = max(0, int(round(value)))
-        hours, remainder = divmod(total, 3600)
-        minutes, seconds = divmod(remainder, 60)
-        if hours:
-            return f"{hours:02d}:{minutes:02d}:{seconds:02d}"
-        return f"{minutes:02d}:{seconds:02d}"
-
-    # ------------------------------------------------------------------ Watch loop -----------------------------------------------------------
-    def _start_watch_loop(self) -> None:
-        if self._closed or self._watcher or self._watch_start_pending:
-            return
-        self._watch_start_pending = True
-        threading.Thread(target=self._watch_start_worker, daemon=True).start()
-
-    def _watch_start_worker(self) -> None:
-        try:
-            watcher = mtgo_bridge.start_watch(interval_ms=self.WATCH_INTERVAL_MS)
-        except FileNotFoundError as exc:
-            logger.error("Bridge executable not found: {}", exc)
-            if not self._closed:
-                wx.CallAfter(self._handle_watch_start_failure, "timer.status.bridge_missing", exc)
-            return
-        except Exception as exc:  # noqa: BLE001
-            logger.exception("Unable to start bridge watcher")
-            if not self._closed:
-                wx.CallAfter(self._handle_watch_start_failure, "timer.status.bridge_error", exc)
-            return
-
-        if self._closed:
-            self._stop_watcher_worker(watcher)
-            return
-        wx.CallAfter(self._complete_watch_start, watcher)
-
-    def _handle_watch_start_failure(self, status_key: str, error: Exception) -> None:
-        self._watch_start_pending = False
-        if self._closed:
-            return
-        if status_key == "timer.status.bridge_missing":
-            self._set_status(status_key)
-        else:
-            self._set_status(status_key, error=error)
-        wx.CallLater(self.WATCH_RETRY_DELAY_MS, self._start_watch_loop)
-
-    def _complete_watch_start(self, watcher: BridgeWatcher) -> None:
-        self._watch_start_pending = False
-        if self._closed:
-            self._stop_watcher_async(watcher)
-            return
-        if self._watcher is not None and self._watcher is not watcher:
-            self._stop_watcher_async(watcher)
-            return
-        self._watcher = watcher
-        self._watch_timer.Start(self.WATCH_INTERVAL_MS)
-
-    def _stop_watcher_async(self, watcher: BridgeWatcher | None = None) -> None:
-        watcher_to_stop = watcher or self._watcher
-        if watcher is None:
-            self._watcher = None
-        if watcher_to_stop is None:
-            return
-        threading.Thread(
-            target=self._stop_watcher_worker,
-            args=(watcher_to_stop,),
-            daemon=True,
-        ).start()
-
-    def _stop_watcher_worker(self, watcher: BridgeWatcher) -> None:
-        try:
-            watcher.stop()
-        except Exception as exc:  # noqa: BLE001
-            logger.debug(f"Failed to stop bridge watcher: {exc}")
-
-    def _on_watch_timer(self, _event: wx.TimerEvent) -> None:
-        if not self._watcher:
-            return
-        payload = self._watcher.latest()
-        if not payload:
-            return
-        self._last_snapshot = payload
-        self._update_challenge_display(payload)
-
-    def _on_monitor_timer(self, _event: wx.TimerEvent) -> None:
-        self._monitor_timer_step()
-
-    # ------------------------------------------------------------------ Helpers --------------------------------------------------------------
-    def _set_status(self, key: str, **kwargs: object) -> None:
-        self.status_text.ChangeValue(self._t(key, **kwargs))
-
-    def _update_challenge_display(self, snapshot: dict[str, Any]) -> None:
-        if not self.challenge_text:
-            return
-        timers = snapshot.get("challengeTimers") or []
-        if not timers:
-            self.challenge_text.SetLabel(self._t("timer.no_challenge"))
-            self.challenge_text.Wrap(self._challenge_wrap_width())
-            return
-        lines: list[str] = []
-        for timer in timers:
-            desc = timer.get("description") or timer.get("eventId") or "Challenge"
-            state = timer.get("state") or "Unknown"
-            remaining = timer.get("remainingSeconds")
-            remaining_display = self._format_seconds(remaining)
-            fmt = timer.get("format")
-            line = desc
-            if fmt and fmt.lower() != "no format found":
-                line += f" • {fmt}"
-            line += f" ({state}) — {remaining_display}"
-            lines.append(line)
-        self.challenge_text.SetLabel("\n".join(lines))
-        self.challenge_text.Wrap(self._challenge_wrap_width())
-
-    def _challenge_wrap_width(self) -> int:
-        if not self.challenge_text:
-            return 340
-        parent = self.challenge_text.GetParent()
-        width = parent.GetClientSize().width if parent else 0
-        return max(240, width - 24)
-
-    def _on_resize(self, event: wx.Event) -> None:
-        if self.challenge_text:
-            self.challenge_text.Wrap(self._challenge_wrap_width())
-        event.Skip()
-
-    # ------------------------------------------------------------------ Lifecycle -------------------------------------------------------------
-    def on_close(self, event: wx.CloseEvent) -> None:
-        self._closed = True
-        self._watch_start_pending = False
-        if self._watch_timer.IsRunning():
-            self._watch_timer.Stop()
-        if self._monitor_timer.IsRunning():
-            self._monitor_timer.Stop()
-        if self._repeat_timer.IsRunning():
-            self._repeat_timer.Stop()
-        self._stop_watcher_async()
-        event.Skip()
-
 
 def main() -> None:
     """Launch the timer alert widget as a standalone application."""
@@ -638,6 +355,3 @@ def main() -> None:
 
 if __name__ == "__main__":
     main()
-
-
-__all__ = ["TimerAlertFrame"]
